// src/utils/enums/envConfig.enum.ts
export enum EnvConfigEnum {
  APP_AWS_REGION = 'APP_AWS_REGION',
  APP_AWS_ACCESS_KEY_ID = 'APP_AWS_ACCESS_KEY_ID',
  APP_AWS_SECRET_ACCESS_KEY = 'APP_AWS_SECRET_ACCESS_KEY',
  APP_AWS_BUCKET = 'APP_AWS_BUCKET',
  PORT = 'PORT',
  TOKEN_SECRET = 'TOKEN_SECRET',
  REFRESH_TOKEN_SECRET = 'REFRESH_TOKEN_SECRET',
  CONNECTION_STRING = 'CONNECTION_STRING',
  TOKEN_EXPIRATION_TIME = 'TOKEN_EXPIRATION_TIME',
  REFRESH_TOKEN_EXPIRATION_TIME = 'REFRESH_TOKEN_EXPIRATION_TIME',
  RESET_PASSWORD_TOKEN_EXPIRATION_TIME = 'RESET_PASSWORD_TOKEN_EXPIRATION_TIME',
  GOOGLE_CLIENT_ID = 'GOOGLE_CLIENT_ID',
  GOOGLE_CLIENT_SECRET = 'GOOGLE_CLIENT_SECRET',
  GOOGLE_URL = 'GOOGLE_URL',
  GOOGLE_REDIRECT_URL = 'GOOGLE_REDIRECT_URL',
  REDIRECT_URI= 'REDIRECT_URI',
  FACEBOOK_CLIENT_ID= 'FACEBOOK_CLIENT_ID',
  FACEBOOK_CLIENT_SECRET= 'FACEBOOK_CLIENT_SECRET',
  FACEBOOK_REDIRECT_URI= 'FACEBOOK_REDIRECT_URI',
  MAILGUN_DOMAIN= 'MAILGUN_DOMAIN',
  NITRALS_EMAIL_ADDRESS= 'NITRALS_EMAIL_ADDRESS',
  MAILGUN_API_KEY= "MAILGUN_API_KEY",
  MAILGUN_HOST= 'MAILGUN_HOST',
  FACEBOOK_REDIRECT_URL= 'FACEBOOK_REDIRECT_URL',
  REMITA_BASE_URL='REMITA_BASE_URL',
  REMITA_API_KEY='REMITA_API_KEY',
  REMITA_SECRET_KEY= 'REMITA_SECRET_KEY'
}